* 'master' of github.com:kevadsett/ZombieCity:
  updated building prefabs for updated models & changed the way ui hooks up so we don't have prefab cross dependency (yay statics!)
  added the toppanel ui prefab, refactored the daynight control
  split hud up into sub prefabs so we conflict less :) :)  :)
  placeholder looting ui that looks too pretty to be placeholder
  high rise building and wonkified low rise
  day night controller
  city spawning tweaks so zombies are inside buildings far less frequently
  fix
  proper house now

+++ Unity/Assets/Scripts/World/Building.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Building : MonoBehaviour {
 
-	public int clearRadius;
+	public float clearRadius;
 	public float minScale;
 	public float maxScale;
